requirements.txt
fastapi
uvicorn[standard]
pydantic
spacy
python-dotenv
requests
